fix: Column type in PheWAS model

+++ app/queries/models/phewas.py
@@ -6,7 +6,7 @@ class PheWAS(db.Model):
     __tablename__ = 'phewas'
 
     id = db.Column(db.BigInteger, primary_key=True)
-    gwas_id_n = db.Column(db.Integer)
+    gwas_id_n = db.Column(db.BigInteger)
     snp_id = db.Column(db.String(255))
     chr_id = db.Column(db.SmallInteger, nullable=False)
     pos = db.Column(db.Integer, nullable=False)
